Extract the selection classes into a separate file
